Move express into express_app and import it into server_part.js
