+++ webpack.config.js
@@ -1,9 +1,9 @@
 const path = require('path');
 
 module.exports = {
-  entry: './src/index.js',
+  entry: './public/frontend.js',
   output: {
-    path: path.resolve(__dirname, 'public'),
-    filename: 'script_with_bundle_experiment.js',
+    filename: 'bundle.js',
+    path: path.resolve(__dirname, 'dist'),
   },
 };
